+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.beta1 20??-??-??
+ - 2015-07-13 Updated CPAN module MIME::Tools to version 5.506.
  - 2015-07-13 Updated CPAN module Locale::Codes to version 3.35.
  - 2015-07-13 Updated CPAN module Linux::Distribution to version 0.23.
  - 2015-07-13 Updated CPAN module libwww-perl to version 6.13.